requirements.txt
amqp==5.0.5
asgiref==3.2.10
billiard==3.6.3.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1
django-debug-toolbar==3.2
django-phone-field==1.8.1
django-phonenumber-field==5.0.0
kombu==5.0.2
phonenumbers==8.12.17
prompt-toolkit==3.0.14
psycopg2==2.8.6
python-dotenv==0.15.0
pytz==2021.1
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
